backend/requirements.txt
Django==5.1.15
djangorestframework==3.15.2
django-filter==24.3
psycopg2-binary==2.9.11
django-environ==0.11.2
django-cors-headers==4.6.0
stripe==11.4.1
requests==2.32.3
celery[redis]==5.4.0
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0
gunicorn==23.0.0
